Update paramConfig.json
 estimate for number of hospitalized was slightly too high compared to Finland situation.  Change: -> 2,1%
ICU ratio: average in Finland duering the last one month: 23%
length of the hospitalisation: 13 days is very close to stats reported by THL. I would not change.
IFR of 0,60% is on a bit high compared to consensus estimates, but the results produced by the model and in combination with the undetected infections, this is not too low (could be even 20% higher)
ratio of undetected infections. Now 83%. This is tricky... but if we change this to higher (?) we would need to tune IFR and hosp rate (lower). @tuomas.syrjanen , @Heikki Lukkarinen comments :)
Lenght of incubation period: 5,2 days. This seems to be close to consensus (e.g. WHO: 5-6 days) but could be a but lower (see the following): 4,0?
T inf (time the patient is infected) is more tricky, because the model does not differentiate between start of symptoms and infectiousness. This significantly impacts the serial interval and should propably be shorter than the original default of 2,9 (->1,5?). The serial interval is typically estimated to be 4-5 (or up to 6,5 days). Also: I can not reproduce the speed of the spread in March if the serial interval is e.g. 8.